# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChittyOS GPT Connector API",
    "description": "Comprehensive API for connecting custom GPTs to ChittyOS services, third-party integrations, and banking connections",
    "version": "1.0.0",
    "contact": {
      "name": "ChittyOS Support",
      "url": "https://chitty.cc"
    }
  },
  "servers": [
    {
      "url": "https://connect.chitty.cc",
      "description": "Production ChittyConnect API"
    },
    {
      "url": "https://connect-staging.chitty.cc",
      "description": "Staging environment"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/api/thirdparty/notion/query": {
      "post": {
        "summary": "Query Notion Database",
        "description": "Query Notion database via ChittyConnect proxy",
        "operationId": "queryNotion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "databaseId": {
                    "type": "string"
                  },
                  "filter": {
                    "type": "object"
                  },
                  "sorts": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                },
                "required": ["databaseId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/thirdparty/neon/query": {
      "post": {
        "summary": "Query Neon Database",
        "description": "Execute SQL query on Neon database",
        "operationId": "queryNeon",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  },
                  "params": {
                    "type": "array",
                    "items": {}
                  }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query executed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "rowCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ChittyOS-API-Key"
      }
    },
...
  }
}